Users can now add tasks to their lists
Using the textbox at the top of the listwindow, users can type the title
of the task they want to add, however currently they cannot add a description.
I've had to make some changes to the structure of the program, but I've
tried to keep the more functional approach instead of using mutable
references everywhere.